Added Spotless and User AuthenticationService

+++ core/src/main/java/io/j13n/core/commons/security/jwt/ContextUser.java
@@ -1,18 +1,18 @@
 package io.j13n.core.commons.security.jwt;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Data;
+import lombok.ToString;
+import lombok.experimental.Accessors;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+
 import java.io.Serial;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import lombok.Data;
-import lombok.ToString;
-import lombok.experimental.Accessors;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 @Data
 @Accessors(chain = true)
@@ -22,9 +22,9 @@ public class ContextUser implements Serializable {
     @Serial
     private static final long serialVersionUID = 3430294045916132045L;
 
-    private BigInteger id;
-    private BigInteger createdBy;
-    private BigInteger updatedBy;
+    private Long id;
+    private Long createdBy;
+    private Long updatedBy;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private String userName;
